% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatrap.R
\name{aquatrap}
\alias{aquatrap}
\title{Aquamean}
\usage{
aquatrap(ENV, quant = c(0.01, 0.25, 0.75, 0.99))
}
\arguments{
\item{ENV}{A data frame with each column as an environmental variable}

\item{quant}{A numeric vector with 4 quarantines. Default is .01, 0.25, 0.75, 0.99 quantiles.}
}
\value{
A myAquaOut object 1) trap.env - trapezoid shaped environmental variables, 2) env - imputed environmental data, 3) metrics - a data frame with trapezoid parameters for each environmental variable
}
\description{
Aquamean
}
\examples{
x <-which(test_data$occurance==1)
ENV<-test_data[x,4:6]
aquatrap(ENV, quant = c(.01, 0.25, 0.75, 0.99))
myAquaOut <- aquatrap(ENV, quant = c(.01, 0.25, 0.75, 0.99))
}
