% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatrap_pred.R
\name{aquatrap_pred}
\alias{aquatrap_pred}
\title{A function to predict the TSI index to a new set (or future) environemntal variables}
\usage{
aquatrap_pred(ENV_new, myAquaOut, na.rm = FALSE)
}
\arguments{
\item{ENV_new}{A set of new environmental data to be used for prediction}

\item{myAquaOut}{output of aquatrap function}

\item{na.rm}{A logical indicating whether missing values should be removed when calculating the niche. Default is FALSE}
}
\value{
a list object with 2 elements: trap.env - a trapeze environmental variables and yul- HSI index (probability of occurance)
}
\description{
A function to predict the TSI index to a new set (or future) environemntal variables
}
\examples{
x <- which(test_data$occurance==1)
ENV<-test_data[x , 4:6]
myAquaOut <- aquatrap(ENV, quant = c(.01, 0.25, 0.75, 0.99))
ENVnew<-test_data[ ,4:6]
prde_newENV <- aquatrap_pred(ENVnew, myAquaOut)
}
