% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Example of Occurrence data.}
\format{
A data frame with 259200 rows and 6 variables:
\describe{
  \item{Lon}{longitude}
  \item{Lat}{lstitude}
  \item{occurance}{presence data}
  \item{SST}{sea surface tempertaure in degrees (C)}
  \item{SSS}{sea surface salinity in (ppm)}
  \item{NPP}{net primamry production (g C m−2 yr−1)}
  ...
}
}
\source{
\url{https://obis.org/}
}
\usage{
test_data
}
\description{
A dataset containing the swordfish presence data, 1x1 degrees coordinate grid and set of environemntal variables (temperature, salinity, primamry productivity)
}
\keyword{datasets}
